Moved connections to seperate file
